fix(privacy): the lead card's task pair reads one task, and the offer spans the whole thread
Five review findings.

The lead card showed one task's title beside another task's deadline. The
subject subquery took the audience and the due_at subquery beside it did
not, so on a lead whose next open task is limited they picked different
rows under the same ORDER BY. The two describe ONE task and now select one
row; my earlier comment claiming the sibling "owes nothing" was wrong the
moment they had to agree.

The due-threads offer filtered `captured_by LIKE 'connector:%'` before
testing the audience, so a HAND-LOGGED limited message on a captured
thread was invisible to it while the window read — which is not
connector-scoped — could see it. No limited text reached the model (the
window excludes it), but the pass would summarise a conversation while one
of its messages is withheld from the summary's readers, with nothing
saying a part is missing. The test is a correlated subquery over every
email on the thread, spelled as NOT EXISTS over the limited ones rather
than bool_and over the open ones: the two differ on a thread whose
messages are all somehow absent, and only this direction refuses it.

`SetCaptureLabel` also guards `restricted_at IS NULL`. The trigger already
refuses a label arriving after a hold; this covers the other order, which
commits the hold with a derived label on the row that the trigger then
refuses every attempt to remove.

The census accepted `audience IS NOT NULL` as a guard. The column is NOT
NULL, so that is a predicate every row passes — `IS` is dropped and only a
comparison against a value counts.

The raw-capture fixture used fixed `source_id` values under a UNIQUE
(source_system, source_id); a rerun against a database this suite did not
rebuild would have failed on a duplicate key rather than an assertion. Per
run now.

The signal-extraction suite crossed the 1000-line test ceiling, so the
audience cases move to signalextractaudience_integration_test.go: what the
pass may READ, apart from what it concludes. A new fixture there seeded
against now() while the suite freezes its clock — the fixture-clock gate
caught it, and the timestamp is derived from the frozen value.
